Extract common caching functionality to common package
